refactor: ToastContext
