--- src/generic/toast-context/index.js
@@ -1,3 +1,0 @@
-// @ts-check
-export { default as ToastContext } from './context';
-export { default as useToastContext } from './hooks';
